Making the plugin work with $().clone(true)
The plugin was originally built using closures to implement the event
handlers and accessed variables from enclosing scope to reference
the element jquery object. Thus, when a cloned element was created, and
received events, the event handlers worked on the cloning source, when they
should work on the cloned element.
Only works with jquery 1.4+ as it relies on 1.4 version feature of cloning
not only events, but data as well.