Fixing browser-lockup issue
When a masked field has focus, a user might click out of the browser
window.  Then when a user clicks back into that input field (before the
browser has focus), a recursive loop seems to lockup the browser and
crash it.

This fix was submitted by Bryan Drenner to me to submit online here.