Clear buffer after input blur
When you type in the middle of the mask (ie. "(___) __2-____" for a phone number) if the mask is optional, the number will jump to the beginning of the mask on blur (ie. the input value becomes "(2" ) Then when you click back in the input the buffer will add the 2 back in the position it was in when you blurred the input (ie. the input value becomes "(2__) __2-____" ) This makes it necessary to clear the buffer on input blur